refactor(analysis): extract cyclomatic complexity module
